Description
This file is the powerhouse of the article page. It made the call to the backend API and transformed all the data into usable content. The api came in a JSON format. This service did most of the work so that each page, module, or component didn't have to. It also limited calls to the backend versus having each page, module, or component do it independently.