Description
This is the main article page. It housed unique AI created articles. There were several components that made up the page such as an image and video carousel, the main article, recommended articles that related to the page, and a list of trending articles. It 
also housed a sidekick widget which was basically a container for an ad. This page was fully responsive from a wide screen > 1250px 
down to a mobile size of 300px.