Description
This is the trending section component. It gave a list of trending articles that related to the main article. It was fully responsive and also had a bottomless scroll feature so that if the user scrolled down to a certain point then it would load 10 more items to the list. It was designed to be completely seamless to the user. It was called bottomless however there was a limit of 100 list items if I remember correctly. If it hit that number or it reached to end of trending articles to show then the scroll feature would turn off and the user would see the footer. Also, depending on screen width, there were ad widgets placed every nth position within the list.